Refactoring
Deleting and replacing redundant files.
