+++ src/rendering/config/FileDescriptorSetBlueprint.java
@@ -1,8 +1,7 @@
-package demos.helloDescriptor.rendering;
+package rendering.config;
 
 import static rendering.pipeline.PipelineUtil.*;
 
-import java.nio.LongBuffer;
 import java.util.List;
 import java.util.logging.Level;
 import java.util.logging.Logger;
@@ -15,11 +14,12 @@ import core.resources.ConfigAsset;
 import core.resources.ConfigFile;
 import core.resources.Destroyable;
 import core.result.VulkanException;
-import rendering.pipeline.ShaderBlueprint;
+import rendering.engine.shader.DescriptorBlueprint;
+import rendering.engine.shader.DescriptorSetBlueprint;
 
-public class FileShaderBlueprint implements ShaderBlueprint, Destroyable {
+public class FileDescriptorSetBlueprint implements DescriptorSetBlueprint, Destroyable {
 	
-	private final static Logger logger = Logger.getLogger(FileShaderBlueprint.class.getName());
+	private final static Logger logger = Logger.getLogger(FileDescriptorSetBlueprint.class.getName());
 	
 	private static final String 
 		BINDINGS_COUNT_KEY = "bindingsCount",
@@ -37,21 +37,22 @@ public class FileShaderBlueprint implements ShaderBlueprint, Destroyable {
 	
 	private VkDevice device;
 	private long descriptorLayout;
+	private int bindings;
 	VkDescriptorSetLayoutBinding.Buffer layoutBindings;
 	
 	
-	public FileShaderBlueprint(VkDevice device, ConfigFile cfg) {
+	public FileDescriptorSetBlueprint(VkDevice device, ConfigFile cfg) {
 		this.device = device;
 		loadFromAsset(cfg);
 		cfg.close();
 	}
 	
 	private void loadFromAsset(ConfigAsset cfg) {
-		layoutBindings = VkDescriptorSetLayoutBinding.calloc(cfg.getInteger(BINDINGS_COUNT_KEY, DEFAULT_BINDINGS_COUNT));
+		bindings = cfg.getInteger(BINDINGS_COUNT_KEY, DEFAULT_BINDINGS_COUNT);
+		layoutBindings = VkDescriptorSetLayoutBinding.calloc(bindings);
 		List<ConfigAsset> cfgs = cfg.getCfgList(BINDINGS_KEY);
 		
 		for (int i = 0; i < cfgs.size(); i++) {
-			
 			try {
 				layoutBindings.get(i)
 					.binding(cfgs.get(i).getInteger(BINDING_KEY, DEFAULT_BINDING))
@@ -73,14 +74,19 @@ public class FileShaderBlueprint implements ShaderBlueprint, Destroyable {
 	}
 	
 	@Override
-	public LongBuffer createLayoutsBuffer(VkDevice device) {
-		return null;
+	public long getLayout() {
+		return descriptorLayout;
 	}
 
 	@Override
 	public void destroy() {
 		layoutBindings.free();
 		VK10.vkDestroyDescriptorSetLayout(device, descriptorLayout, null);
+	}
+	
+	@Override
+	public int descriptorCount() {
+		return bindings;
 	}
 
 }
