--- src/rendering/pipeline/Shader.java
@@ -1,7 +1,0 @@
-package rendering.pipeline;
-
-public class Shader {
-
-	//TODO: recordable, descriptor value updates
-	
-}
