--- src/rendering/pipeline/ShaderBlueprint.java
@@ -1,11 +1,0 @@
-package rendering.pipeline;
-
-import java.nio.LongBuffer;
-
-import org.lwjgl.vulkan.VkDevice;
-
-public interface ShaderBlueprint {
-
-	public LongBuffer createLayoutsBuffer(VkDevice device);
-	
-}
